Added count and average calculations and other

+++ dotNetTask.API/Controllers/EmployeeController.cs
@@ -22,15 +22,19 @@ namespace dotNetTask.API.Controllers
             _mapper = mapper;
         }
 
+        //GET api/employee/{id}
         [HttpGet("{id}")]
         public async Task<ActionResult<EmployeeDto>> GetEmployeeAsync(Guid id)
         {
-            var employeeForRepository = await _employeeRepository.GetEmployeeAsync(id);
-            return Ok(_mapper.Map<EmployeeDto>(employeeForRepository));
+            var employeeFromRepository = await _employeeRepository.GetEmployeeAsync(id);
+            if (employeeFromRepository is null) return NotFound();
+            return Ok(_mapper.Map<EmployeeDto>(employeeFromRepository));
             
         }
-        //pagal intervala    
+        //pagal intervala
 
+
+        //GET api/employee
         [HttpGet]
         public async Task<ActionResult<IEnumerable<EmployeeDto>>> GetEmployeesAsync()
         {
@@ -40,6 +44,7 @@ namespace dotNetTask.API.Controllers
             return Ok(_mapper.Map<IEnumerable<EmployeeDto>>(employeesFromRepository));
         }
 
+        // GET api/employee/ByBossId/{bossId}
         [HttpGet("ByBossId/{bossId}")]
         public async Task<ActionResult<IEnumerable<EmployeeDto>>> GetEmployeesByBossIdAsync(Guid bossId)
         {
@@ -47,18 +52,25 @@ namespace dotNetTask.API.Controllers
             return Ok(_mapper.Map<IEnumerable<EmployeeDto>>(employeeFromrepository));
         }
 
+        // GET api/employee/CountAndAverage/{role}
         [HttpGet("/CountAndAvarage/{role}")]
-        public async Task<ActionResult<IEnumerable<EmployeeDto>>> GetCauntAndAvarageByRoleAsync(string role)
+        public async Task<ActionResult<CountAndAverage>> GetCauntAndAvarageByRoleAsync(string role)
         {
-            var employeeFromrepository = await _employeeRepository.GetCauntAndAvarageByRoleAsync(role);
-            return Ok(_mapper.Map<IEnumerable<EmployeeDto>>(employeeFromrepository));
-            
+            var countAndAverageByRole = await _employeeRepository.GetCauntAndAvarageByRoleAsync(role);
+
+            return Ok(countAndAverageByRole);
         }
 
         // POST api/employee
         [HttpPost]
         public async Task<ActionResult<Employee>> AddEmployeeAsync(CreateEmployeeDto employeeDto)
         {
+            if (employeeDto.Role == EmployeeRoles.CEO && await _employeeRepository.CheckIsCeoExistAsync()) return BadRequest("CEO exist");
+
+            Employee getBoss = null;
+            if (employeeDto.Role != EmployeeRoles.CEO) getBoss = await _employeeRepository.GetEmployeeAsync(employeeDto.BossId);
+            
+            
             Employee employeeToCreate = new()
             {
                 Id = Guid.NewGuid(),
@@ -66,39 +78,50 @@ namespace dotNetTask.API.Controllers
                 LastName = employeeDto.LastName,
                 BirtDate = employeeDto.BirtDate,
                 EmploymentDate = employeeDto.EmploymentDate,
-                Boss = null,
+                Boss = getBoss,
                 HomeAddress = employeeDto.HomeAddress,
                 CurrentSalary = employeeDto.CurrentSalary,
                 Role = employeeDto.Role
             };
+
             var createdEmployee = await _employeeRepository.AddNewEmployeeAsync(employeeToCreate);
-            var createdEmployeeToReturn = _mapper.Map<EmployeeDto>(createdEmployee);
-            return Ok(createdEmployeeToReturn);
+
+            return CreatedAtAction(nameof(GetEmployeeAsync), new{ id = employeeToCreate.Id }, employeeToCreate);
         }
 
-        // PUT api/employee/
+        // PUT api/employee/{id}
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateEmployeeAsync(Guid id, UpdateEmployeeDto employeeDto)
         {
             var existingEmployee = await _employeeRepository.GetEmployeeAsync(id);
+            
             if (existingEmployee is null) return NotFound();
            
             var updatedEmployee = _mapper.Map(employeeDto, existingEmployee);
+
             await _employeeRepository.UpdateEmployeeAsync(updatedEmployee);
-            return Ok();
+
+            return NoContent();
         }
 
+        // PATCH api/employee/salary
         [HttpPatch]
         public ActionResult UpdateEmployeeSalary()
         {
             return Ok();
         }
 
+        // DELETE api/employee/{id}
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteEmployeeAsync(Guid id)
         {
+            var existingEmployee = await _employeeRepository.GetEmployeeAsync(id);
+            
+            if (existingEmployee is null) return NotFound();
+
             await _employeeRepository.DeleteEmployee(id);
-            return Ok();
+
+            return NoContent();
         }
     }
 }
